Basic issue support (#8)
* Add extended support for fetching issue fields from JIRA, some other
functionality was added to support this such as deserializing additional
JIRA objects. A full list is below:

  * Added basic OAuth support, an access token can now be passed
    it is up to the implementing app to do the OAuth dance and
    provide scope.

  * Split out history, changelogs and issue_types into their own files
    for a smaller and more focused issue.rs

  * Added an API call for getting issue_types by id.

  * Added an API call for getting resolutions by id.

  * Added more supported fields to the issue struct.

* Add support for various structures within issues:

  * Add ability to get attachments and have attachments represented
    in the issue struct.

  * Add ability to get projects and have projects represented in
    the issue struct.

  * Add watcher and time_tracking structs to issue.

+++ src/v2/application_role.rs
@@ -52,15 +52,15 @@ pub struct ApplicationRole {
 
     /// Number of seats
     #[serde(rename = "numberOfSeats", default)]
-    pub number_of_seats: usize,
+    pub number_of_seats: i64,
 
     /// Remaining seats
     #[serde(rename = "remainingSeats", default)]
-    pub remaining_seats: usize,
+    pub remaining_seats: i64,
 
     /// User count
     #[serde(rename = "userCount", default)]
-    pub user_count: usize,
+    pub user_count: i64,
 
     /// User count description
     #[serde(rename = "userCountDescription", default)]
